examples/requirements_whatsapp.txt
# Dépendances pour le bot WhatsApp CSS
# Installation: pip install -r requirements_whatsapp.txt

# Framework web pour les webhooks
Flask==2.3.3
Werkzeug==2.3.7

# Client HTTP pour les appels API
requests==2.31.0

# Gestion des variables d'environnement
python-dotenv==1.0.0

# Cache et sessions (optionnel)
redis==5.0.1

# Validation des données
pydantic==2.5.0

# Logging avancé
loguru==0.7.2

# Gestion des dates
python-dateutil==2.8.2

# Serveur WSGI pour production
gunicorn==21.2.0

# Monitoring et métriques (optionnel)
prometheus-client==0.19.0

# Sécurité
cryptography==41.0.7

# Utilitaires
click==8.1.7
colorama==0.4.6